feat: run sync cli
