+++ src/utils/time.ts
@@ -6,3 +6,13 @@ export function getSecsToMidnight(): number {
 
   return Math.floor((tomorrow.getTime() - now.getTime()) / 1000)
 }
+
+export function parseTime(timeString: string): Date {
+  const [hours, minutes, seconds] = timeString.split(":").map(Number)
+  const date = new Date()
+  date.setHours(hours ?? date.getHours())
+  date.setMinutes(minutes ?? date.getMinutes())
+  date.setSeconds(seconds ?? date.getSeconds())
+
+  return date
+}
